+++ client/src/ts/account.ts
@@ -3,8 +3,10 @@ import '../scss/account.scss';
 import topNavbar from './modules/global/topNavbar';
 import botNavbar from './modules/global/botNavbar';
 import { initAccount } from './modules/account/initAccount';
+import { disableBackForwardCache } from './modules/global/disableBackForwardCache';
 
-// initializing imports
+disableBackForwardCache();
+
 topNavbar();
 botNavbar();