+++ client/src/ts/index.ts
@@ -2,7 +2,9 @@ import '../scss/index.scss';
 
 import topNavbar from './modules/global/topNavbar';
 import botNavbar from './modules/global/botNavbar';
+import { disableBackForwardCache } from './modules/global/disableBackForwardCache';
 
-// initializing imports
+disableBackForwardCache();
+
 topNavbar();
 botNavbar();