+++ client/src/ts/modules/accountRecovery/recoveryEmailForm.ts
@@ -81,7 +81,6 @@ async function startAccountRecovery(e: SubmitEvent): Promise<void> {
     const asyncErrorData: AsyncErrorData | null = getAsyncErrorData(err);
 
     if (!asyncErrorData) {
-      popup('Something went wrong.', 'error');
       return;
     };