+++ client/src/ts/modules/accountRecovery/recoveryPasswordUpdateForm.ts
@@ -84,7 +84,6 @@ async function updateAccountPassword(e: SubmitEvent): Promise<void> {
     const asyncErrorData: AsyncErrorData | null = getAsyncErrorData(err);
 
     if (!asyncErrorData) {
-      popup('Something went wrong.', 'error');
       return;
     };
 
@@ -179,7 +178,6 @@ async function resendAccountRecoveryEmail(): Promise<void> {
     const asyncErrorData: AsyncErrorData | null = getAsyncErrorData(err);
 
     if (!asyncErrorData) {
-      popup('Something went wrong.', 'error');
       return;
     };