+++ client/src/ts/modules/global/errorUtils.ts
@@ -1,4 +1,6 @@
 import axios, { AxiosError } from "../../../../node_modules/axios/index";
+import { InfoModal } from "./InfoModal";
+import popup from "./popup";
 
 export interface AsyncErrorData {
   status: number,
@@ -9,12 +11,19 @@ export interface AsyncErrorData {
 
 export function getAsyncErrorData(err: unknown): AsyncErrorData | null {
   if (!axios.isAxiosError(err)) {
+    popup('Something went wrong.', 'error');
     return null;
   };
 
   const axiosError: AxiosError<AxiosErrorResponseData> = err;
 
   if (!axiosError.status || !axiosError.response) {
+    popup('Something went wrong.', 'error');
+    return null;
+  };
+
+  if (axiosError.status === 429) {
+    handleRateLimitReached();
     return null;
   };
 
@@ -24,4 +33,14 @@ export function getAsyncErrorData(err: unknown): AsyncErrorData | null {
     errReason: axiosError.response.data.reason,
     errResData: axiosError.response.data.resData,
   };
+};
+
+function handleRateLimitReached(): void {
+  popup('Too many requests.', 'error');
+
+  InfoModal.display({
+    title: 'Please slow down.',
+    description: `You're making too many requests. Please wait 30 to 60 seconds to be able to make further requests.`,
+    btnTitle: 'Okay',
+  }, { simple: true });
 };