+++ client/src/ts/modules/hangout/chat/hangoutChat.ts
@@ -223,7 +223,6 @@ async function getHangoutMessages(): Promise<void> {
     const asyncErrorData: AsyncErrorData | null = getAsyncErrorData(err);
 
     if (!asyncErrorData) {
-      popup('Something went wrong.', 'error');
       return;
     };
 
@@ -280,7 +279,6 @@ async function sendHangoutMessage(e: SubmitEvent): Promise<void> {
     const asyncErrorData: AsyncErrorData | null = getAsyncErrorData(err);
 
     if (!asyncErrorData) {
-      popup('Something went wrong.', 'error');
       return;
     };