+++ client/src/ts/modules/hangout/dashboard/handleNotHangoutMember.ts
@@ -111,7 +111,6 @@ export async function joinHangoutAsAccount(): Promise<void> {
     const asyncErrorData: AsyncErrorData | null = getAsyncErrorData(err);
 
     if (!asyncErrorData) {
-      popup('Something went wrong.', 'error');
       return;
     };