+++ client/src/ts/modules/hangout/dashboard/hangoutDashboard.ts
@@ -103,7 +103,6 @@ export async function getInitialHangoutData(): Promise<void> {
     const asyncErrorData: AsyncErrorData | null = getAsyncErrorData(err);
 
     if (!asyncErrorData) {
-      popup('Something went wrong.', 'error');
       return;
     };
 
@@ -452,7 +451,6 @@ async function leaveHangout(): Promise<void> {
     const asyncErrorData: AsyncErrorData | null = getAsyncErrorData(err);
 
     if (!asyncErrorData) {
-      popup('Something went wrong.', 'error');
       return;
     };