+++ client/src/ts/modules/hangout/dashboard/hangoutDashboardUtils.ts
@@ -102,7 +102,6 @@ export async function handleNotSignedIn(hangoutId: string): Promise<void> {
     const asyncErrorData: AsyncErrorData | null = getAsyncErrorData(err);
 
     if (!asyncErrorData) {
-      popup('Something went wrong.', 'error');
       return;
     };