+++ client/src/ts/modules/hangout/members/hangoutMembers.ts
@@ -271,7 +271,6 @@ async function transferHangoutLeadership(newLeaderMemberId: number): Promise<voi
     const asyncErrorData: AsyncErrorData | null = getAsyncErrorData(err);
 
     if (!asyncErrorData) {
-      popup('Something went wrong.', 'error');
       return;
     };
 
@@ -377,7 +376,6 @@ async function kickHangoutMember(memberToKickId: number): Promise<void> {
     const asyncErrorData: AsyncErrorData | null = getAsyncErrorData(err);
 
     if (!asyncErrorData) {
-      popup('Something went wrong.', 'error');
       return;
     };
 
@@ -463,7 +461,6 @@ async function relinquishHangoutLeadership(): Promise<void> {
     const asyncErrorData: AsyncErrorData | null = getAsyncErrorData(err);
 
     if (!asyncErrorData) {
-      popup('Something went wrong.', 'error');
       return;
     };
 
@@ -552,7 +549,6 @@ async function claimHangoutLeadership(): Promise<void> {
     const asyncErrorData: AsyncErrorData | null = getAsyncErrorData(err);
 
     if (!asyncErrorData) {
-      popup('Something went wrong.', 'error');
       return;
     };